BigQueryテストフローの改善と修正
- index.phpでのクラス読み込み問題を修正
- AdminerLoginBigQueryとAdminerBigQueryCSSの名前空間問題を解決
- BigQuery認証、データセット表示、テーブル構造表示、SELECT実行機能の動作確認完了
- 11データセット、346,368行のページング機能動作確認
- 日本語テキスト表示の正常動作確認
- Fatal/Error/Invalid警告なしでの安定動作確認
