+++ devtools/issues/i06.md
@@ -31,7 +31,7 @@
 
 # 3. クラス分離
 * bigquery.php に含まれているクラスをファイルに分割します。
-* 1クラスつづ作業します。今回はAdminerLoginBigQueryクラスを分離します。
+* 1クラスつづ作業します。今回はAdminerBigQueryCSSクラスを分離します。
 * 分離されたファイルはplugins/drivers/bigquery/に配置してください
 
 
@@ -40,3 +40,18 @@
 * SQLを受け取って実行するメソッドを適切なクラス内に作成し、グローバルfunctionではSQL文をそのメソッドに渡すだけにしてください。
 * adminer/drivers/mysql.inc.php を参考にしてください。
 * bigquery.php内のすべてのfunctionについて、同様の冗長コード排除を行ってください
+
+# 5. クラス分離
+* bigquery.php に含まれているクラスをファイルに分割します。
+* 1クラスつづ作業します。今回はBigQueryUtilsクラスを分離します。
+* 分離されたファイルはplugins/drivers/bigquery/に配置してください
+
+# 6. クラス分離
+* bigquery.php に含まれているクラスをファイルに分割します。
+* 今回はResult, Driverクラスを分離します。
+* 分離されたファイルはplugins/drivers/bigquery/に配置してください
+
+# 7. クラス分離
+* bigquery.php に含まれているクラスをファイルに分割します。
+* 今回はDBクラスを分離します。
+* 分離されたファイルはplugins/drivers/bigquery/に配置してください
