+++ devtools/web/index.php
@@ -14,7 +14,7 @@ function adminer_object()
 	require_once __DIR__ . '/plugins/drivers/bigquery.php';
 
 	$plugins = array(
-		new AdminerLoginBigQuery(array(
+		new \Adminer\AdminerLoginBigQuery(array(
 			'project_id' => getenv('GOOGLE_CLOUD_PROJECT')
 		)),
 		new \Adminer\AdminerBigQueryCSS(),
@@ -25,7 +25,7 @@ function adminer_object()
 
 // OAuth2コールバック処理（Adminer実行前に処理）
 $is_oauth_callback = isset($_GET['oauth2']) && $_GET['oauth2'] === 'callback' &&
-                    isset($_GET['code']) && isset($_GET['state']);
+	isset($_GET['code']) && isset($_GET['state']);
 
 if ($is_oauth_callback) {
 	// デバッグログ出力
